pyproject.toml
[project]
name = "notice-api"
version = "0.0.0"
description = ""
authors = [
    { name = "alan910127", email = "70696274+alan910127@users.noreply.github.com" },
]
dependencies = [
    "fastapi>=0.105.0",
    "uvicorn[standard]>=0.24.0.post1",
    "pydantic>=2.5.2",
    "pydantic-settings>=2.1.0",
    "loguru>=0.7.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }


[project.optional-dependencies]
dump = ["click>=8.1.7", "PyYAML>=6.0.1"]
[tool.pdm.dev-dependencies]
dev = ["pyright>=1.1.339", "ruff>=0.1.7"]

[tool.pdm.scripts]
dev = "uvicorn notice_api.main:app --reload --port 8000 --host 0.0.0.0"
format = "ruff format src"
lint = "ruff check src"
typecheck = "pyright src"
dump-spec = "python -m notice_api.dump_spec"

[tool.ruff.isort]
known-first-party = ["notice_api"]

[tool.ruff]
extend-select = [
    "W",   # pycodestyle warnings
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "PGH", # pygrep-hooks
    "RUF", # ruff
]
